# iterate on baskets + an optimization