--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -683,10 +683,10 @@ end
     @test all(onesrow .== 1)
 end
 
-@testset "basketarrays()" begin
+@testset "basketarray_iter()" begin
     f = UnROOT.samplefile("tree_with_vector_multiple_baskets.root")
     t = LazyTree(f,"t1")
-    @test (UnROOT.basketarrays(f, f["t1"]["b1"]) .|> length) == [1228, 1228, 44]
-    @test (UnROOT.basketarrays(t.b1) .|> length) == [1228, 1228, 44]
+    @test (UnROOT.basketarray_iter(f, f["t1"]["b1"]) .|> length) == [1228, 1228, 44]
+    @test (UnROOT.basketarray_iter(t.b1) .|> length) == [1228, 1228, 44]
     @test length(UnROOT.basketarray(t.b1, 1)) == 1228
 end