Add some functions to iterate on baskets.

This lets one bypass the cost of many `getindex()`s...

```julia
julia> @time sum(t.nMuon) |> Int
  0.716353 seconds (6.31 k allocations: 470.014 MiB, 3.63% gc time)
149322456

julia> @time sum(sum,UnROOT.basketarrays(t.nMuon)) |> Int
  0.215638 seconds (6.71 k allocations: 470.039 MiB, 13.26% gc time)
149322456
```
well, also highlighting that `getindex()` can be improved @Moelf :) Also note that uproot takes 0.5s to do the above and C++ root takes 1.1s.


By the way,
```julia
julia> @profilehtml sum(t.nMuon)
```
shows a gap after the orange getindex bar. Not sure what that means. I would've expected the bars above getindex to add up to the full orange bar width...

![image](https://user-images.githubusercontent.com/5760027/133739149-5d127f3d-f6a5-4557-8f3d-0566320f58c7.png)