@Moelf 2021-09-17
>getindex() can be improved\

yeah indeed but I doubt it can ever be improved 3x. Basically when one iterates over events, especially in parallel, we have a branch:
https://github.com/tamasgal/UnROOT.jl/blob/ac66158f38d8b0fecec06595473f9879b3f693d8/src/iteration.jl#L151

that we can't avoid... I think it's performance would look relatively better in the case where the physics workload is >> a few branches...