@aminnj 2021-09-17
Latest commit factors out the uncommon/slow path from getindex(). This evidently helps the compiler because now we cut down on the getindex() time by a factor of almost 2. Tested on two machines and I see similar reductions.

```julia
julia> function bar(t)
           for evt in t
               _ = evt.Muon_pt
           end
           nothing
       end
```

### before
(note this is on a different machine wrt the original post, hence the faster times to begin with)
```julia
julia> @time bar(t)
  0.888482 seconds (14.30 k allocations: 1.818 GiB, 4.61% gc time)

julia> @time sum(t.nMuon) |> Int
  0.495250 seconds (6.08 k allocations: 469.980 MiB, 6.55% gc time)
149322456
```

### after
```julia
julia> @time bar(t)
  0.762924 seconds (14.30 k allocations: 1.818 GiB, 5.33% gc time)

julia> @time sum(t.nMuon) |> Int
  0.224379 seconds (6.08 k allocations: 469.980 MiB, 5.02% gc time)
149322456
```