@aminnj 2021-09-17
After the change, the gap is more or less gone.
![image](https://user-images.githubusercontent.com/5760027/133847171-85f0b10c-9a14-45e9-8a3b-ca1f667a7eda.png)