@Moelf 2021-09-17
wow, insane!! doing lord's work thanks this is great.

so now all the time are spent in copying, on the 
https://github.com/tamasgal/UnROOT.jl/blob/ac66158f38d8b0fecec06595473f9879b3f693d8/src/root.jl#L272