@aminnj 2021-09-17
Well to put it into perspective all these gains probably get diluted by a factor of 3 when we run on the zlib file, but better than nothing :). 

We can get blazing fast speeds if we eliminate that pesky copy! 👿