@Moelf 2021-09-17
btw I think
>basketarrays
basketarray

is getting out of hands