@aminnj 2021-09-17
We could change it to `basketarray_iter`. what's the Julia convention?