@Moelf 2021-09-17
there's not really convention, I just don't know how people are gonna use it. `_iter` is fine, one word is also fine.