@aminnj 2021-09-17
I'll rename it to `_iter` since then it's more clear what it does. We're not exporting any of the `basketarray*` methods so these are mainly in existence if someone knows what they're doing and want high performance (operating on chunks of a branch)